MacOS spaceline theme fix
The wave theme as has problem with color due to a bug in XPM image
generation. See here for description. http://archive.is/qN56q

Thus, we will use wave them in MacOS only if we are using the Emacs
mac port, otherwise, we shall use utf separators for spaceline.