Paths should be set before other packages
Other packages might require all the system paths to be set, in order
to properly select the available program on the system. For example,
ispell will choose wrong program if the path is not set correctly.

This can be set by making sure that exec-path-from-shell is loaded as
soon as possible.