Use prog-mode hooks
prog-mode is a meta mode in Emacs that is supposed to be derived by
each programming mode. Unfortunately, MATLAB and CC modes don't derive
from prog-mode. In this case, we explicitly invoke the prog-mode hook
for these modes. This further implies that we don't have to invoke
common things like line numbers individually for these modes.

In future, a common package can be added to prog-mode-hook and will be
automatically available in these modes too.