Pin magit to Melpa Stable
In recent version, Magit has changed how [Enter] works in `diff`
buffers.

In the stable version, if a file is open, and its diff is
being viewed in the magit-diff mode, pressing [Enter] will jump to
proper line in the already opened file frame. In alpha version of
Magit, the file is alway opened in the `diff` buffer. Why they made
the change, I don't know. But we'll ride it out till they release
their next version.

It might be a good idea to pin all major packages (Helm, Magit,
Use-package, etc) to their stable versions.